+++ panel/util/__init__.py
@@ -22,7 +22,7 @@ from functools import partial
 from html import escape  # noqa
 from importlib import import_module
 from typing import (
-    Any, AnyStr, Dict, Iterable, Iterator, List, Optional, Union,
+    Any, AnyStr, Dict, Iterator, List, Union,
 )
 
 import bokeh
@@ -32,44 +32,15 @@ import param
 from bokeh.model import Model
 from packaging.version import Version
 
-datetime_types = (np.datetime64, dt.datetime, dt.date)
+from .checks import (  # noqa
+    datetime_types, is_dataframe, is_holoviews, is_number, is_parameterized,
+    is_series, isdatetime, isfile, isIn, isurl,
+)
 
 bokeh_version = Version(bokeh.__version__)
 
 
 PARAM_NAME_PATTERN = re.compile(r'^.*\d{5}$')
-
-
-def isfile(path: str) -> bool:
-    """Safe version of os.path.isfile robust to path length issues on Windows"""
-    try:
-        return os.path.isfile(path)
-    except ValueError: # path too long for Windows
-        return False
-
-
-def isurl(obj: Any, formats: Optional[Iterable[str]] = None) -> bool:
-    if not isinstance(obj, str):
-        return False
-    lower_string = obj.lower().split('?')[0].split('#')[0]
-    return (
-        lower_string.startswith('http://')
-        or lower_string.startswith('https://')
-    ) and (formats is None or any(lower_string.endswith('.'+fmt) for fmt in formats))
-
-
-def is_dataframe(obj) -> bool:
-    if 'pandas' not in sys.modules:
-        return False
-    import pandas as pd
-    return isinstance(obj, pd.DataFrame)
-
-
-def is_series(obj) -> bool:
-    if 'pandas' not in sys.modules:
-        return False
-    import pandas as pd
-    return isinstance(obj, pd.Series)
 
 
 def hashable(x):
@@ -79,22 +50,6 @@ def hashable(x):
         return tuple([(k,v) for k,v in x.items()])
     else:
         return x
-
-
-def isIn(obj, objs):
-    """
-    Checks if the object is in the list of objects safely.
-    """
-    for o in objs:
-        if o is obj:
-            return True
-        try:
-            if o == obj:
-                return True
-        except Exception:
-            pass
-    return False
-
 
 def indexOf(obj, objs):
     """
@@ -219,39 +174,6 @@ def get_method_owner(meth):
         return meth.__self__
 
 
-def is_parameterized(obj) -> bool:
-    """
-    Whether an object is a Parameterized class or instance.
-    """
-    return (isinstance(obj, param.Parameterized) or
-            (isinstance(obj, type) and issubclass(obj, param.Parameterized)))
-
-def is_holoviews(obj: Any) -> bool:
-    """
-    Whether the object is a HoloViews type that can be rendered.
-    """
-    if 'holoviews' not in sys.modules:
-        return False
-    from holoviews.core.dimension import Dimensioned
-    from holoviews.plotting import Plot
-    return isinstance(obj, (Dimensioned, Plot))
-
-def isdatetime(value) -> bool:
-    """
-    Whether the array or scalar is recognized datetime type.
-    """
-    if is_series(value) and len(value):
-        return isinstance(value.iloc[0], datetime_types)
-    elif isinstance(value, np.ndarray):
-        return (
-            value.dtype.kind == "M" or
-            (value.dtype.kind == "O" and len(value) != 0 and
-             isinstance(value[0], datetime_types))
-        )
-    elif isinstance(value, list):
-        return all(isinstance(d, datetime_types) for d in value)
-    else:
-        return isinstance(value, datetime_types)
 
 def value_as_datetime(value):
     """
@@ -275,14 +197,6 @@ def datetime_as_utctimestamp(value):
     Converts a datetime to a UTC timestamp used by Bokeh interally.
     """
     return value.replace(tzinfo=dt.timezone.utc).timestamp() * 1000
-
-
-def is_number(s: Any) -> bool:
-    try:
-        float(s)
-        return True
-    except ValueError:
-        return False
 
 
 def parse_query(query: str) -> Dict[str, Any]:
@@ -387,7 +301,7 @@ def lazy_load(module, model, notebook=False, root=None):
             f'\n\npn.extension(\'{ext}\')\n'
         )
     elif root is not None:
-        from .io.state import state
+        from ..io.state import state
         ext = module.split('.')[-1]
         if root.ref['id'] in state._views:
             param.main.param.warning(
